Cleanup before refactoring, hopefully properly this time.
Rewrote the core, this time it looks better.

kimplified the token crawling, hopefully not too much.

This does break the other indent styles, but only temporarily.

Now that things are in better shape, add Allman and GNU with tests

And it's just that easy. Allman, Whitesmiths, Ratliff, Pico and Lisp styles

Added and tested.